Move ServiceTemplateInstanceService to core
