+++ org.opentosca.container.api/src/main/java/org/opentosca/container/api/controller/ServiceTemplateInstanceController.java
@@ -42,9 +42,7 @@ import org.opentosca.container.api.dto.plan.PlanDTO;
 import org.opentosca.container.api.dto.request.CreateServiceTemplateInstanceRequest;
 import org.opentosca.container.api.dto.situations.SituationsMonitorDTO;
 import org.opentosca.container.api.dto.situations.SituationsMonitorListDTO;
-import org.opentosca.container.api.service.InstanceService;
 import org.opentosca.container.api.service.PlanInvokerService;
-import org.opentosca.container.core.next.services.instances.SituationInstanceService;
 import org.opentosca.container.core.common.uri.UriUtil;
 import org.opentosca.container.core.model.csar.Csar;
 import org.opentosca.container.core.next.model.DeploymentTest;
@@ -55,6 +53,8 @@ import org.opentosca.container.core.next.model.ServiceTemplateInstanceState;
 import org.opentosca.container.core.next.model.SituationsMonitor;
 import org.opentosca.container.core.next.repository.ServiceTemplateInstanceRepository;
 import org.opentosca.container.core.next.services.instances.PlanInstanceService;
+import org.opentosca.container.core.next.services.instances.ServiceTemplateInstanceService;
+import org.opentosca.container.core.next.services.instances.SituationInstanceService;
 import org.opentosca.deployment.checks.DeploymentTestService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -67,29 +67,30 @@ public class ServiceTemplateInstanceController {
     private static final Logger logger = LoggerFactory.getLogger(ServiceTemplateInstanceController.class);
     private final Csar csar;
     private final TServiceTemplate serviceTemplate;
-    private final InstanceService instanceService;
     private final SituationInstanceService situationInstanceService;
     private final PlanInstanceService planInstanceService;
     private final PlanInvokerService planInvokerService;
     private final DeploymentTestService deploymentTestService;
     private final ServiceTemplateInstanceRepository serviceTemplateInstanceRepository;
+    private final ServiceTemplateInstanceService serviceTemplateInstanceService;
     @Context
     private UriInfo uriInfo;
 
     public ServiceTemplateInstanceController(final Csar csar, final TServiceTemplate serviceTemplate,
-                                             final InstanceService instanceService, final PlanInstanceService planInstanceService,
+                                             final PlanInstanceService planInstanceService,
                                              final PlanInvokerService planInvokerService,
                                              final DeploymentTestService deploymentTestService,
                                              final SituationInstanceService situationInstanceService,
-                                             ServiceTemplateInstanceRepository serviceTemplateInstanceRepository) {
+                                             final ServiceTemplateInstanceRepository serviceTemplateInstanceRepository,
+                                             final ServiceTemplateInstanceService serviceTemplateInstanceService) {
         this.csar = csar;
         this.serviceTemplate = serviceTemplate;
-        this.instanceService = instanceService;
         this.planInstanceService = planInstanceService;
         this.planInvokerService = planInvokerService;
         this.deploymentTestService = deploymentTestService;
         this.situationInstanceService = situationInstanceService;
         this.serviceTemplateInstanceRepository = serviceTemplateInstanceRepository;
+        this.serviceTemplateInstanceService = serviceTemplateInstanceService;
     }
 
     @GET
@@ -98,7 +99,7 @@ public class ServiceTemplateInstanceController {
     public Response getServiceTemplateInstances() {
         logger.debug("Invoking getServiceTemplateInstances");
         final Collection<ServiceTemplateInstance> serviceInstances =
-            this.instanceService.getServiceTemplateInstances(serviceTemplate.getId());
+            this.serviceTemplateInstanceService.getServiceTemplateInstances(serviceTemplate.getId());
         logger.debug("Found <{}> instances of ServiceTemplate \"{}\" ", serviceInstances.size(), serviceTemplate.getId());
 
         final ServiceTemplateInstanceListDTO list = new ServiceTemplateInstanceListDTO();
@@ -127,7 +128,7 @@ public class ServiceTemplateInstanceController {
 
         try {
             final ServiceTemplateInstance createdInstance =
-                this.instanceService.createServiceTemplateInstance(csar.id().csarName(), serviceTemplate.getId(), request.getCorrelationId().trim());
+                this.serviceTemplateInstanceService.createServiceTemplateInstance(csar.id().csarName(), serviceTemplate.getId(), request.getCorrelationId().trim());
 
             final URI uri = UriUtil.generateSubResourceURI(this.uriInfo, createdInstance.getId().toString(), false);
 
@@ -196,7 +197,7 @@ public class ServiceTemplateInstanceController {
     @ApiOperation(hidden = true, value = "")
     public Response deleteServiceTemplateInstance(@PathParam("id") final Long id) {
         logger.debug("Invoking deleteServiceTemplateInstance");
-        this.instanceService.deleteServiceTemplateInstance(id);
+        this.serviceTemplateInstanceService.deleteServiceTemplateInstance(id);
         return Response.noContent().build();
     }
 
@@ -213,7 +214,7 @@ public class ServiceTemplateInstanceController {
     @ApiOperation(value = "Get state of a service template instance", response = String.class)
     public Response getServiceTemplateInstanceState(@ApiParam("ID of service template instance") @PathParam("id") final Long id) {
         logger.debug("Invoking getServiceTemplateInstanceState");
-        final ServiceTemplateInstanceState state = this.instanceService.getServiceTemplateInstanceState(id);
+        final ServiceTemplateInstanceState state = this.serviceTemplateInstanceService.getServiceTemplateInstanceState(id);
         return Response.ok(state.toString()).build();
     }
 
@@ -224,7 +225,7 @@ public class ServiceTemplateInstanceController {
     public Response updateServiceTemplateInstanceState(@PathParam("id") final Long id, final String request) {
         logger.debug("Invoking updateServiceTemplateInstanceState");
         try {
-            this.instanceService.setServiceTemplateInstanceState(id, request);
+            this.serviceTemplateInstanceService.setServiceTemplateInstanceState(id, request);
         } catch (final IllegalArgumentException e) { // this handles a null request too
             return Response.status(Status.BAD_REQUEST).build();
         }
@@ -237,7 +238,7 @@ public class ServiceTemplateInstanceController {
     @ApiOperation(hidden = true, value = "")
     public Response getServiceTemplateInstanceProperties(@PathParam("id") final Long id) {
         logger.debug("Invoking getServiceTemplateInstanceProperties");
-        final ServiceTemplateInstance instance = this.instanceService.getServiceTemplateInstance(id, true);
+        final ServiceTemplateInstance instance = this.serviceTemplateInstanceService.getServiceTemplateInstance(id, true);
         final Document properties = instance.getPropertiesAsDocument();
 
         if (properties == null) {
@@ -253,7 +254,7 @@ public class ServiceTemplateInstanceController {
     @ApiOperation(value = "Gets the properties of a service template instance", response = Map.class)
     public Map<String, String> getServiceTemplateInstancePropertiesAsJSON(@PathParam("id") final Long id) {
         logger.debug("Invoking getServiceTemplateInstancePropertiesAsJSON");
-        final ServiceTemplateInstance serviceTemplateInstance = this.instanceService.getServiceTemplateInstance(id, true);
+        final ServiceTemplateInstance serviceTemplateInstance = this.serviceTemplateInstanceService.getServiceTemplateInstance(id, true);
         return serviceTemplateInstance.getPropertiesAsMap();
     }
 
@@ -265,7 +266,7 @@ public class ServiceTemplateInstanceController {
     public Response updateServiceTemplateInstanceProperties(@PathParam("id") final Long id, final Document request) {
         logger.debug("Invoking updateServiceTemplateInstanceProperties");
         try {
-            this.instanceService.setServiceTemplateInstanceProperties(id, request);
+            this.serviceTemplateInstanceService.setServiceTemplateInstanceProperties(id, request);
         } catch (final IllegalArgumentException e) { // this handles a null request too
             return Response.status(Status.BAD_REQUEST).build();
         } catch (final ReflectiveOperationException e) {
@@ -295,7 +296,7 @@ public class ServiceTemplateInstanceController {
     @Produces( {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
     public Response createSituationMonitor(@PathParam("id") final Long id, SituationsMonitorDTO monitor) {
         logger.debug("Invoking createSituationMonitor");
-        ServiceTemplateInstance servInstance = this.instanceService.getServiceTemplateInstance(id, false);
+        ServiceTemplateInstance servInstance = this.serviceTemplateInstanceService.getServiceTemplateInstance(id, false);
 
         Map<String, Collection<Long>> mapping = new HashMap<>();
 
@@ -317,7 +318,7 @@ public class ServiceTemplateInstanceController {
     private ServiceTemplateInstance resolveInstance(final Long instanceId, final String templateId) throws NotFoundException {
         // We only need to check that the instance belongs to the template, the rest is
         // guaranteed while this is a sub-resource
-        final ServiceTemplateInstance instance = this.instanceService.getServiceTemplateInstance(instanceId, false);
+        final ServiceTemplateInstance instance = this.serviceTemplateInstanceService.getServiceTemplateInstance(instanceId, false);
 
         if (!instance.getTemplateId().equals(templateId)) {
             logger.info("Service template instance <{}> could not be found", instanceId);
