+++ org.opentosca.container.api/src/main/java/org/opentosca/container/api/service/InstanceService.java
@@ -7,48 +7,37 @@ import java.util.Optional;
 import javax.inject.Inject;
 import javax.ws.rs.NotFoundException;
 import javax.xml.namespace.QName;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
 
 import org.eclipse.winery.model.tosca.TRelationshipTemplate;
 import org.eclipse.winery.model.tosca.TServiceTemplate;
 
 import org.opentosca.container.api.dto.request.CreateRelationshipTemplateInstanceRequest;
-import org.opentosca.container.api.util.Utils;
 import org.opentosca.container.core.common.jpa.DocumentConverter;
 import org.opentosca.container.core.engine.ToscaEngine;
+import org.opentosca.container.core.model.ModelUtils;
 import org.opentosca.container.core.model.csar.Csar;
 import org.opentosca.container.core.model.csar.CsarId;
 import org.opentosca.container.core.next.model.NodeTemplateInstance;
 import org.opentosca.container.core.next.model.NodeTemplateInstanceProperty;
 import org.opentosca.container.core.next.model.NodeTemplateInstanceState;
-import org.opentosca.container.core.next.model.PlanInstance;
-import org.opentosca.container.core.next.model.PlanType;
 import org.opentosca.container.core.next.model.Property;
 import org.opentosca.container.core.next.model.RelationshipTemplateInstance;
 import org.opentosca.container.core.next.model.RelationshipTemplateInstanceProperty;
 import org.opentosca.container.core.next.model.RelationshipTemplateInstanceState;
-import org.opentosca.container.core.next.model.ServiceTemplateInstance;
-import org.opentosca.container.core.next.model.ServiceTemplateInstanceProperty;
-import org.opentosca.container.core.next.model.ServiceTemplateInstanceState;
 import org.opentosca.container.core.next.repository.NodeTemplateInstanceRepository;
-import org.opentosca.container.core.next.repository.PlanInstanceRepository;
 import org.opentosca.container.core.next.repository.RelationshipTemplateInstanceRepository;
 import org.opentosca.container.core.next.repository.ServiceTemplateInstanceRepository;
-import org.opentosca.container.core.next.services.instances.PlanInstanceService;
 import org.opentosca.container.core.next.services.templates.RelationshipTemplateService;
-import org.opentosca.container.core.next.services.templates.ServiceTemplateService;
 import org.opentosca.container.core.service.CsarStorageService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.w3c.dom.Document;
-import org.w3c.dom.Element;
 
 /**
  * Allows access to instance information for service templates and node templates.
  */
+// TODO: split node template and relationship template handling in separate repos
 @Service
 public class InstanceService {
 
@@ -59,10 +48,7 @@ public class InstanceService {
     private final RelationshipTemplateInstanceRepository relationshipTemplateInstanceRepository;
 
     private final RelationshipTemplateService relationshipTemplateService;
-    private final ServiceTemplateService serviceTemplateService;
     private final CsarStorageService storage;
-    private final PlanInstanceRepository planInstanceRepository;
-    private final PlanInstanceService planInstanceService;
 
     private final DocumentConverter converter = new DocumentConverter();
 
@@ -71,195 +57,16 @@ public class InstanceService {
                            NodeTemplateInstanceRepository nodeTemplateInstanceRepository,
                            RelationshipTemplateInstanceRepository relationshipTemplateInstanceRepository,
                            RelationshipTemplateService relationshipTemplateService,
-                           ServiceTemplateService serviceTemplateService,
-                           CsarStorageService storage, PlanInstanceRepository planInstanceRepository, PlanInstanceService planInstanceService) {
+                           CsarStorageService storage) {
         this.serviceTemplateInstanceRepository = serviceTemplateInstanceRepository;
         this.nodeTemplateInstanceRepository = nodeTemplateInstanceRepository;
         this.relationshipTemplateInstanceRepository = relationshipTemplateInstanceRepository;
         this.relationshipTemplateService = relationshipTemplateService;
-        this.serviceTemplateService = serviceTemplateService;
         this.storage = storage;
-        this.planInstanceRepository = planInstanceRepository;
-        this.planInstanceService = planInstanceService;
     }
 
     public Document convertPropertyToDocument(final Property property) {
         return this.converter.convertToEntityAttribute(property.getValue());
-    }
-
-    /* Service Template Instances */
-    public Collection<ServiceTemplateInstance> getServiceTemplateInstances(final String serviceTemplate) {
-        logger.debug("Requesting instances of ServiceTemplate \"{}\"...", serviceTemplate);
-        return this.serviceTemplateInstanceRepository.findByTemplateId(serviceTemplate);
-    }
-
-    public ServiceTemplateInstance getServiceTemplateInstance(final Long id, final boolean evaluatePropertyMappings) {
-        logger.debug("Requesting service template instance <{}>...", id);
-        final Optional<ServiceTemplateInstance> instance = this.serviceTemplateInstanceRepository.findWithNodeAndRelationshipTemplateInstancesById(id);
-
-        if (instance.isPresent()) {
-            final ServiceTemplateInstance result = instance.get();
-
-            if (evaluatePropertyMappings) {
-                final PropertyMappingsHelper helper = new PropertyMappingsHelper(this, storage);
-                helper.evaluatePropertyMappings(result);
-            }
-
-            return result;
-        }
-
-        logger.debug("Service Template Instance <" + id + "> not found.");
-        throw new NotFoundException("Service Template Instance <" + id + "> not found.");
-    }
-
-    public ServiceTemplateInstanceState getServiceTemplateInstanceState(final Long id) {
-        final ServiceTemplateInstance service = getServiceTemplateInstance(id, false);
-
-        return service.getState();
-    }
-
-    public void setServiceTemplateInstanceState(final Long id, final String state) throws NotFoundException,
-        IllegalArgumentException {
-
-        ServiceTemplateInstanceState newState;
-        try {
-            newState = ServiceTemplateInstanceState.valueOf(state);
-        } catch (final Exception e) {
-            final String msg =
-                String.format("The given state %s is an illegal service template instance state.", state);
-            logger.error(msg, e);
-            throw new IllegalArgumentException(msg, e);
-        }
-
-        final ServiceTemplateInstance service = getServiceTemplateInstance(id, false);
-        service.setState(newState);
-        this.serviceTemplateInstanceRepository.save(service);
-    }
-
-    public void setServiceTemplateInstanceProperties(final Long id,
-                                                     final Document properties) throws ReflectiveOperationException {
-        final ServiceTemplateInstance service = getServiceTemplateInstance(id, false);
-
-        try {
-            final ServiceTemplateInstanceProperty property =
-                Utils.convertDocumentToProperty(properties, ServiceTemplateInstanceProperty.class);
-            service.addProperty(property);
-            this.serviceTemplateInstanceRepository.save(service);
-        } catch (InstantiationException | IllegalAccessException e) { // This is not supposed to happen at all!
-            final String msg = String.format("An error occurred while instantiating an instance of the %s class.",
-                ServiceTemplateInstanceProperty.class);
-            logger.error(msg, e);
-            throw e;
-        }
-    }
-
-    public void deleteServiceTemplateInstance(final Long instanceId) {
-        // throws exception if not found
-        final ServiceTemplateInstance instance = getServiceTemplateInstance(instanceId, false);
-        this.serviceTemplateInstanceRepository.delete(instance);
-    }
-
-    public ServiceTemplateInstance createServiceTemplateInstance(final String csarId, final String serviceTemplateName) throws InstantiationException, IllegalAccessException, IllegalArgumentException {
-        final CsarId csar = this.serviceTemplateService.checkServiceTemplateExistence(csarId, serviceTemplateName);
-        final Document propertiesAsDoc =
-            createServiceInstanceInitialPropertiesFromServiceTemplate(csar, serviceTemplateName);
-        final ServiceTemplateInstanceProperty property =
-            Utils.convertDocumentToProperty(propertiesAsDoc, ServiceTemplateInstanceProperty.class);
-
-        final ServiceTemplateInstance instance = new ServiceTemplateInstance();
-        instance.setCsarId(csar);
-        instance.setTemplateId(serviceTemplateName);
-        instance.setState(ServiceTemplateInstanceState.INITIAL);
-        instance.addProperty(property);
-
-        this.serviceTemplateInstanceRepository.save(instance);
-
-        return instance;
-    }
-
-    public ServiceTemplateInstance createServiceTemplateInstance(final String csarId, final String serviceTemplateName,
-                                                                 final String correlationId) throws NotFoundException,
-        InstantiationException,
-        IllegalAccessException,
-        IllegalArgumentException {
-        final CsarId csar = this.serviceTemplateService.checkServiceTemplateExistence(csarId, serviceTemplateName);
-
-        PlanInstance pi = (PlanInstance) this.planInstanceService.waitForInstanceAvailable(correlationId).joinAndGet(30000);
-
-        // if no instance was found it is possible that live-modeling was started, just create an empty instance
-        if (pi == null) {
-            return this.createServiceTemplateInstance(csarId, serviceTemplateName);
-        }
-
-        // If the found plan is a build plan there shouldn't be a service template instance available,
-        // if it is a transformation plan the service instance mustn't be of the service template the new service instance should belong to
-        if ((pi.getType().equals(PlanType.BUILD) && pi.getServiceTemplateInstance() == null)
-            || (pi.getType().equals(PlanType.TRANSFORMATION) && !pi.getServiceTemplateInstance().getTemplateId().equals(serviceTemplateName))) {
-
-            return this.createServiceTemplateInstance(csar, serviceTemplateName, pi);
-        } else {
-            final String msg = "The build plan instance is already associted with a service template instance!";
-            logger.error(msg);
-            throw new IllegalArgumentException(msg);
-        }
-    }
-
-    private ServiceTemplateInstance createServiceTemplateInstance(final CsarId csarId, final String serviceTemplateName,
-                                                                  final PlanInstance buildPlanInstance) throws InstantiationException,
-        IllegalAccessException,
-        IllegalArgumentException {
-        final Document propertiesAsDoc =
-            createServiceInstanceInitialPropertiesFromServiceTemplate(csarId, serviceTemplateName);
-        final ServiceTemplateInstanceProperty property =
-            Utils.convertDocumentToProperty(propertiesAsDoc, ServiceTemplateInstanceProperty.class);
-
-        ServiceTemplateInstance instance = new ServiceTemplateInstance();
-        instance.setCsarId(csarId);
-        instance.setTemplateId(serviceTemplateName);
-        instance.setState(ServiceTemplateInstanceState.INITIAL);
-        instance.addProperty(property);
-        instance.addPlanInstance(buildPlanInstance);
-        instance.setCreationCorrelationId(buildPlanInstance.getCorrelationId());
-
-        instance = this.serviceTemplateInstanceRepository.save(instance);
-
-        if (buildPlanInstance.getServiceTemplateInstance() == null) {
-            buildPlanInstance.setServiceTemplateInstance(instance);
-        }
-        planInstanceRepository.save(buildPlanInstance);
-
-        return instance;
-    }
-
-    private Document createServiceInstanceInitialPropertiesFromServiceTemplate(final CsarId csarId,
-                                                                               final String serviceTemplateId) {
-
-        final Document existingProperties =
-            this.serviceTemplateService.getPropertiesOfServiceTemplate(csarId, serviceTemplateId);
-
-        if (existingProperties != null) {
-            return existingProperties;
-        }
-
-        logger.debug("No Properties found in BoundaryDefinitions for ST {} thus creating blank ones",
-            serviceTemplateId);
-        final DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
-        dbf.setNamespaceAware(true);
-        DocumentBuilder db;
-        try {
-            db = dbf.newDocumentBuilder();
-            final Document doc = db.newDocument();
-            final Element createElementNS =
-                doc.createElementNS("http://docs.oasis-open.org/tosca/ns/2011/12", "Properties");
-            createElementNS.setAttribute("xmlns:tosca", "http://docs.oasis-open.org/tosca/ns/2011/12");
-            createElementNS.setPrefix("tosca");
-            doc.appendChild(createElementNS);
-
-            return doc;
-        } catch (final ParserConfigurationException e) {
-            logger.error("Cannot create a new DocumentBuilder: {}", e.getMessage());
-            return null;
-        }
     }
 
     /* Node Template Instances */
@@ -354,7 +161,7 @@ public class InstanceService {
 
         try {
             final NodeTemplateInstanceProperty property =
-                Utils.convertDocumentToProperty(properties, NodeTemplateInstanceProperty.class);
+                ModelUtils.convertDocumentToProperty(properties, NodeTemplateInstanceProperty.class);
             node.addProperty(property);
             this.nodeTemplateInstanceRepository.save(node);
         } catch (InstantiationException | IllegalAccessException e) { // This is not supposed to happen at all!
@@ -463,7 +270,7 @@ public class InstanceService {
 
         try {
             final RelationshipTemplateInstanceProperty property =
-                Utils.convertDocumentToProperty(properties, RelationshipTemplateInstanceProperty.class);
+                ModelUtils.convertDocumentToProperty(properties, RelationshipTemplateInstanceProperty.class);
             relationship.addProperty(property);
             this.relationshipTemplateInstanceRepository.save(relationship);
         } catch (InstantiationException | IllegalAccessException e) { // This is not supposed to happen at all!
@@ -509,7 +316,7 @@ public class InstanceService {
 
         if (propertiesAsDocument != null) {
             final RelationshipTemplateInstanceProperty properties =
-                Utils.convertDocumentToProperty(propertiesAsDocument, RelationshipTemplateInstanceProperty.class);
+                ModelUtils.convertDocumentToProperty(propertiesAsDocument, RelationshipTemplateInstanceProperty.class);
             newInstance.addProperty(properties);
         }
         // State
