+++ org.opentosca.container.core/src/main/java/org/opentosca/container/core/model/ModelUtils.java
@@ -47,9 +47,11 @@ import org.eclipse.winery.model.tosca.TServiceTemplate;
 
 import com.google.common.collect.Sets;
 import org.opentosca.container.core.common.NotFoundException;
+import org.opentosca.container.core.common.jpa.DocumentConverter;
 import org.opentosca.container.core.convention.Types;
 import org.opentosca.container.core.engine.ToscaEngine;
 import org.opentosca.container.core.model.csar.Csar;
+import org.opentosca.container.core.next.model.Property;
 import org.opentosca.container.core.next.xml.PropertyParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -66,6 +68,29 @@ public abstract class ModelUtils {
     public static String makeValidNCName(final String string) {
         return string.replaceAll("\\.", "_").replaceAll(" ", "_").replace("{", "_").replace("}", "_").replace("/", "_")
             .replace(":", "_");
+    }
+
+    /**
+     * Converts an xml document to an xml-based property sui/table for service or node template instances
+     */
+    public static <T extends Property> T convertDocumentToProperty(final Document propertyDoc,
+                                                                   final Class<T> type) throws InstantiationException,
+        IllegalAccessException,
+        IllegalArgumentException {
+
+        if (propertyDoc == null) {
+            final String msg =
+                String.format("The set of parameters of an instance of type %s cannot be null", type.getName());
+            LOG.error(msg);
+            throw new IllegalArgumentException(msg);
+        }
+        final String propertyAsString = new DocumentConverter().convertToDatabaseColumn(propertyDoc);
+        final T property = type.newInstance();
+        property.setName("xml");
+        property.setType("xml");
+        property.setValue(propertyAsString);
+
+        return property;
     }
 
     public static TOperation findNodeOperation(Csar csar, String interfaceName, String operationName) {
